Platform specific backend system now defines what GameSystem is used. Android and Desktop use the standard GameSystem, web uses WebGameSystem. WebGameSystem implements the game-loop using requestAnimationFrame. Updated teavm to use 0.5.1, requires more development.

+++ src/com/linxonline/mallet/io/formats/json/web/WebJSONObject.java
@@ -79,7 +79,7 @@ public class WebJSONObject extends JSONObject
 	@Override
 	public JSONObject put( final String _key, final long _value )
 	{
-		put( object, _key, _value ) ;
+		put( object, _key, ( double )_value ) ;
 		return this ;
 	}
